requirements.txt
# FilmBot A2A Benchmark - Dependencies
# Install: pip install -r requirements.txt

# LangGraph agent framework
langgraph>=1.1.2
langchain-ollama>=1.0.1
langchain-core>=1.2.19

# A2A protocol
a2a-sdk[http-server]>=0.3.25

# A2A / Marketplace server
uvicorn>=0.37.0
fastapi>=0.115.0

# A2A client
httpx>=0.27.0
